MINOR: add storage_name attribute to map model
The attribute value, if present, will indicate the name for the resource
on the storage and runtime endpoints.

If the attribute doesn't have a value, this means the corresponding
mapfile is loaded into haproxy but resides on the filesystem outside of
the directory the datplane manages files in. As a consequence this map
file will not be able to be manipulated through the dataplane's map
storage and runtime endpoints.